feat: displayed gcode progress in overlay rather than dialog
